demo/requirements.txt
git+https://github.com/KupynOrest/s3od.git
gradio>=4.0.0
fire>=0.5.0
numpy>=1.20.0
Pillow>=9.0.0

